software/pc/requirements.txt
matplotlib >= 3.3.2
pyserial >= 3.4